Typing and docs
