+++ website/docs/localization.mdx
@@ -7,20 +7,20 @@ react-day-picker can be easily localized using the localization packages found i
 
 ## Changing Locale
 
-To change the locale of DayPicker, pass to the [`locale`](./props#locale) prop a [`Locale`](http://date-fns.org/docs/Locale) object.
+To change the locale of DayPicker, pass to the [`locale`](./props#locale) prop a `Locale` object. Refer to the [Locale documentation](http://date-fns.org/docs/Locale) on date-fns website for more details.
 
 For example, to localize the calendar in Spanish:
 
-1. import the Spanish Locale object from dateFns;
+1. import the Spanish Locale object from dateFns
+
+   ;
 2. pass the locale object to the `locale` props.
 
-Refer to the [Locale documentation](http://date-fns.org/docs/Locale) on date-fns website for more information about the `Locale` object.
-
 ```jsx preview
-import esLocale from 'date-fns/locale/es';
+import spanish from 'date-fns/locale/es';
 
 function App() {
-  return <DayPicker locale={esLocale} />;
+  return <DayPicker locale={spanish} />;
 }
 ```
 
@@ -29,20 +29,11 @@ function App() {
 To add right-to-left text direction, set the [`dir`](./props#dir) prop to `rtl`.
 
 ```jsx preview
-import arLocale from 'date-fns/locale/ar-SA';
 import { DayPicker } from 'react-day-picker';
-import 'react-day-picker/lib/style-rtl.css';
+import arabic from 'date-fns/locale/ar-SA';
 
 function App() {
-  const [month, setMonth] = useState(new Date());
-  return (
-    <DayPicker
-      month={month}
-      onMonthChange={setMonth}
-      dir="rtl"
-      locale={arLocale}
-    />
-  );
+  return <DayPicker dir="rtl" locale={arabic} />;
 }
 ```
 
@@ -57,17 +48,13 @@ For example, to switch to hindu-arabic using [`toLocaleString`](https://develope
 - set [`formatWeekNumber`](./props/formatWeekNumber) to format the week numbers;
 
 ```jsx preview
-import React, { useState } from 'react';
+import React from 'react';
+import { format } from 'date-fns';
 import { DayPicker } from 'react-day-picker';
 
-import { format } from 'date-fns';
-import arLocale from 'date-fns/locale/ar-SA';
-
-import 'react-day-picker/lib/style.css';
+import arabic from 'date-fns/locale/ar-SA';
 
 function App() {
-  const [month, setMonth] = useState(new Date());
-
   const nuLocale = 'ar-u-nu-arab';
 
   function formatDay(day) {
@@ -86,14 +73,12 @@ function App() {
 
   return (
     <DayPicker
-      month={month}
-      showWeekNumber
+      locale={arabic}
       dir="rtl"
-      locale={arLocale}
+      showWeekNumber
       formatDay={formatDay}
       formatCaption={formatCaption}
       formatWeekNumber={formatWeekNumber}
-      onMonthChange={setMonth}
     />
   );
 }
