+++ website/package.json
@@ -27,6 +27,7 @@
     "ora": "^4.0.1",
     "prop-types": "^15.7.2",
     "react": "^16.10.2",
+    "react-day-picker": "file:../",
     "react-dom": "^16.10.2",
     "react-frame-component": "^4.1.1"
   },
