+++ website/src/components/PropDetails/prop_types.json
@@ -59,19 +59,9 @@
     },
     "categories": {}
   },
-  "month": {
-    "defaultValue": null,
-    "description": "The month to show in the calendar.",
-    "name": "month",
-    "required": true,
-    "type": {
-      "name": "Date"
-    },
-    "categories": {}
-  },
   "initialMonth": {
     "defaultValue": null,
-    "description": "The month to show in the calendar (controlled component).",
+    "description": "The initial month to show in the calendar.",
     "name": "initialMonth",
     "required": false,
     "type": {
@@ -126,6 +116,16 @@
     "required": true,
     "type": {
       "name": "boolean"
+    },
+    "categories": {}
+  },
+  "month": {
+    "defaultValue": null,
+    "description": "The rendered month. Passing this props will make the component switch to controlled mode: this means you will need to implement `onMonthChange` to enable months navigation.",
+    "name": "month",
+    "required": true,
+    "type": {
+      "name": "Date"
     },
     "categories": {}
   },
